Add lightbox functionality to image gallery
This commit adds the lightbox functionality to the image gallery. When a user clicks on a thumbnail, the image will be displayed in a larger view using a lightbox overlay. The lightbox can be closed by clicking on the 'close' button. The implementation is done in the `js/lightbox.js` file.